feat: simplify inference generation tools
